requirements.txt
discord.py
pymongo